Move stack data struct to separate file
